Integration test suite: modularize `helpers`:
Split into: `file`, `test_harness` and `app_builder`.
