+++ helix-term/tests/test/auto_indent.rs
@@ -1,4 +1,5 @@
-use super::*;
+use crate::test::helpers::{platform_line, test_harness::test_with_config, AppBuilder};
+use indoc::indoc;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn auto_indent_c() -> anyhow::Result<()> {
@@ -6,9 +7,9 @@ async fn auto_indent_c() -> anyhow::Result<()> {
         AppBuilder::default().with_file("foo.c", None),
         // switches to append mode?
         (
-            helpers::platform_line("void foo() {#[|}]#"),
+            platform_line("void foo() {#[|}]#"),
             "i<ret><esc>",
-            helpers::platform_line(indoc! {"\
+            platform_line(indoc! {"\
                 void foo() {
                   #[|\n]#\
                 }
