+++ helix-term/tests/test/prompt.rs
@@ -1,4 +1,4 @@
-use super::*;
+use crate::test::helpers::{test_harness::test_key_sequences, AppBuilder};
 
 #[tokio::test(flavor = "multi_thread")]
 async fn test_history_completion() -> anyhow::Result<()> {
